Convert Monkey Integer values into bytes
Set the ground so values in languages implementing Hermes
can be stored in a byte representation. Currently tested
with Monkey integer values
Also to make things easier integer values in Monkey are now
explicitly declared as 32 bit integers.

+++ Tests/MonkeyTests/evaluator/BuiltinFunctionsTests.swift
@@ -12,11 +12,11 @@ import XCTest
 class BuiltinFunctionsTests: XCTestCase {
     func testLenFunction() throws {
         let tests: [(String, Any)] = [
-            ("len(\"\")", 0),
-            ("len(\"four\")", 4),
-            ("len(\"Hello World\")", 11),
-            ("len([1, 2])", 2),
-            ("len([])", 0),
+            ("len(\"\")", Int32(0)),
+            ("len(\"four\")", Int32(4)),
+            ("len(\"Hello World\")", Int32(11)),
+            ("len([1, 2])", Int32(2)),
+            ("len([])", Int32(0)),
             ("len(1)", "Incorrect argment type expected: String or Array got: Integer at Line: 1, Column: 3"),
             (
                 "len(\"one\", \"two\")",
@@ -27,7 +27,7 @@ class BuiltinFunctionsTests: XCTestCase {
         for test in tests {
             do {
                 let evaluated = try Utils.testEval(input: test.0, environment: Environment())
-                MKAssertInteger(object: evaluated, expected: test.1 as? Int ?? -1)
+                MKAssertInteger(object: evaluated, expected: test.1 as? Int32 ?? -1)
             } catch let error as WrongArgumentCount {
                 XCTAssertEqual(error.description, test.1 as? String ?? "")
             } catch let error as InvalidArgumentType {
@@ -40,7 +40,7 @@ class BuiltinFunctionsTests: XCTestCase {
 
     func testFirstFunction() throws {
         let tests: [(String, Any?)] = [
-            ("first([1, 2])", 1),
+            ("first([1, 2])", Int32(1)),
             ("first([])", nil),
             ("first(1)", "Incorrect argment type expected: Array got: Integer at Line: 1, Column: 5"),
             (
@@ -52,7 +52,7 @@ class BuiltinFunctionsTests: XCTestCase {
         for test in tests {
             do {
                 let evaluated = try Utils.testEval(input: test.0, environment: Environment())
-                if let expected = test.1 as? Int {
+                if let expected = test.1 as? Int32 {
                     MKAssertInteger(object: evaluated, expected: expected)
                 } else {
                     XCTAssert((evaluated == Null.null).value)
@@ -69,7 +69,7 @@ class BuiltinFunctionsTests: XCTestCase {
 
     func testLastFunction() throws {
         let tests: [(String, Any?)] = [
-            ("last([1, 2])", 2),
+            ("last([1, 2])", Int32(2)),
             ("last([])", nil),
             ("last(1)", "Incorrect argment type expected: Array got: Integer at Line: 1, Column: 4"),
             (
@@ -81,7 +81,7 @@ class BuiltinFunctionsTests: XCTestCase {
         for test in tests {
             do {
                 let evaluated = try Utils.testEval(input: test.0, environment: Environment())
-                if let expected = test.1 as? Int {
+                if let expected = test.1 as? Int32 {
                     MKAssertInteger(object: evaluated, expected: expected)
                 } else {
                     XCTAssert((evaluated == Null.null).value)
